More int vs unsigned long long

+++ src/libsodium/crypto_sign/edwards25519sha512batch/ref/sign_edwards25519sha512batch.c
@@ -74,7 +74,7 @@ int crypto_sign_open(
     const unsigned char *pk
     )
 {
-  int i;
+  unsigned long long i;
   unsigned char t1[32], t2[32];
   ge25519 get1, get2, gepk;
   sc25519 schmr, scs;
